Add validation for user existence

+++ ChatAPIProject/ChatAPIProject/Controllers/UserController.cs
@@ -98,6 +98,11 @@ namespace ChatAPIProject.Controllers
         [Route("SendFriendRequest")]
         public IHttpActionResult SendFriendRequest(int recieverId)
         {
+            if (!this.CheckIfUserExist(recieverId))
+            {
+                return this.BadRequest($"User with id {recieverId} does not exist.");
+            }
+
             var userId = this.GetUserId();
             FriendRequestInputModel model = new FriendRequestInputModel
             {
@@ -127,6 +132,11 @@ namespace ChatAPIProject.Controllers
         [Route("AcceptRequest")]
         public IHttpActionResult AcceptRequest(AcceptFriendRequestInputModel model)
         {
+            if (!this.CheckIfUserExist(model.FriendId))
+            {
+                return this.BadRequest($"User with id {model.FriendId} does not exist.");
+            }
+
             var userId = GetUserId();
             var isRequestExist = this.ChechIfRequestExist(userId, STATUS_PENDING, model.FriendId);
             if (!isRequestExist)
@@ -151,6 +161,11 @@ namespace ChatAPIProject.Controllers
         [Route("RejectRequest")]
         public IHttpActionResult RejectRequest(RejectFriendRequestInputModel model)
         {
+            if (!this.CheckIfUserExist(model.FriendId))
+            {
+                return this.BadRequest($"User with id {model.FriendId} does not exist.");
+            }
+
             var userId = GetUserId();
             var isRequestExist = this.ChechIfRequestExist(userId, STATUS_PENDING, model.FriendId);
             if (!isRequestExist)
@@ -174,6 +189,11 @@ namespace ChatAPIProject.Controllers
         [Route("SendMessage")]
         public IHttpActionResult SendMessage(SendMessageInputModel model)
         {
+            if (!this.CheckIfUserExist(model.ReceiverId))
+            {
+                return this.BadRequest($"User with id {model.ReceiverId} does not exist.");
+            }
+
             var userId = this.GetUserId();
             var communication = this.communicationService.GetCommunicationByUsers(userId, model.ReceiverId);
             if(communication == null)
@@ -216,6 +236,10 @@ namespace ChatAPIProject.Controllers
         [Route("deleteFriend")]
         public IHttpActionResult DeleteFriend(int friendId)
         {
+            if (!this.CheckIfUserExist(friendId))
+            {
+                return this.BadRequest($"User with id {friendId} does not exist.");
+            }
             try
             {
               int commId = this.communicationService.DeleteFriend(GetUserId(), friendId);
@@ -266,6 +290,13 @@ namespace ChatAPIProject.Controllers
         {
             var allRequests = this.friendRequestSevice.GetRequests(userId, STATUS_PENDING);
             return allRequests.Any(x => x.FriendId == friendId);
+        }
+
+        private bool CheckIfUserExist(int id)
+        {
+            var user = this.Service.IsExist(id);
+
+            return user != null ? true : false;
         }
     }
 }
